requirements.txt
altair==4.2.0
matplotlib==3.4.3
numpy==1.20.3
pandas==1.3.4
plotly==5.8.0
streamlit==1.10.0
streamlit_option_menu==0.3.2
AgGrid==0.2.3
GridOptionsBuilder==0.2.3